autoApi/requirements.txt
allure_python_commons==2.9.45
Faker==13.15.1
jsonpath==0.82
loguru==0.6.0
pymongo==3.6.0
PyMySQL==1.0.2
pytest==7.0.1
PyYAML==6.0
requests==2.27.1
